+++ src/App.js
@@ -1,14 +1,14 @@
 //import logo from './logo.svg';
 import './App.css';
-import { Login } from './components/login/Login';
+
 import "bootstrap/dist/css/bootstrap.min.css";
-import './Assets/css/Style.css';
+import './pages/login/Style.css';
+import { Login } from './pages/login/Login';
 
 function App() {
   return (
     <div className="App">
     <Login />
-      
     </div>
   );
 }
